Issue #97
- Refactored CollapseWhitespaceTest by removing methods which belonged to ShouldNotProcessResponseTest
- Improved ShouldNotProcessResponseTest
